fix/DTO 정리 및 tenant 폴더로 이동
