--- event/src/main/java/com/pyokemon/event/dto/BookingResponseDto.java
@@ -1,15 +1,0 @@
-package com.pyokemon.event.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class BookingResponseDto {
-  private Long bookingId;
-  private Long eventScheduleId;
-}
