+++ event/src/main/java/com/pyokemon/event/dto/tenant/EventScheduleUpdateDto.java
@@ -1,4 +1,4 @@
-package com.pyokemon.event.dto;
+package com.pyokemon.event.dto.tenant;
 
 import java.time.LocalDateTime;
 import java.util.List;
