+++ event/src/main/java/com/pyokemon/event/service/EventScheduleService.java
@@ -8,8 +8,8 @@ import org.springframework.stereotype.Service;
 
 import com.pyokemon.event.dto.BookingInfoResponseDTO;
 import com.pyokemon.event.dto.EventItemResponseDTO;
-import com.pyokemon.event.dto.EventScheduleDto;
-import com.pyokemon.event.dto.PriceDto;
+import com.pyokemon.event.dto.tenant.EventScheduleDto;
+import com.pyokemon.event.dto.tenant.PriceDto;
 import com.pyokemon.event.dto.PriceWithSeatClassDTO;
 import com.pyokemon.event.dto.SeatInfoResponseDTO;
 import com.pyokemon.event.entity.EventSchedule;
